+++ verifyMerkleProofBinary.js
